+++ web-wallet/src/lib/components/__tests__/ContractOperations.spec.js
@@ -21,19 +21,19 @@ describe("ContractOperations", () => {
       {
         disabled: true,
         id: "stake",
-        label: "stake",
+        label: "Stake",
         primary: true,
       },
       {
         disabled: false,
-        id: "unstake",
-        label: "unstake",
+        id: "withdraw-stake",
+        label: "Withdraw Stake",
         primary: false,
       },
       {
         disabled: false,
         id: "claim-rewards",
-        label: "claim rewards",
+        label: "Claim Rewards",
         primary: false,
       },
     ],